Create To-Do List Website
I have crafted a To-Do List Website using Html, CSS, JavaScript. In this Website you can add the tasks and you can prioritize them like low, medium or high and there is a calendar to add the date and you can delete the tasks and also can mark the tasks as completed.